refactor: 検索フォームと結果表示をSearchForm.tsxに統合し、機能単位で完結する構成に変更
